Don't link logic to client id. Use registered player id for that.

+++ UnityProject/Assets/Scripts/Data/PlayerButtonClickData.cs
@@ -2,7 +2,7 @@ namespace Victorina
 {
     public class PlayerButtonClickData
     {
-        public ulong PlayerId { get; set; }
+        public byte PlayerId { get; set; }
         public string Name { get; set; }
         public float Time { get; set; }
     }
